task3's subtask a and b review comments resolved
