+++ app/components/filter-menu.js
@@ -1,0 +1,10 @@
+import Component from "@ember/component";
+
+export default Component.extend({
+  actions: {
+    touchEnd: function () {
+      this.toggleProperty("isShowingBody");
+      this.toggleProperty("clicked");
+    },
+  },
+});
