--- app/components/filter-nav.js
@@ -1,11 +1,0 @@
-import Component from '@ember/component';
-
-export default Component.extend({
-  actions: {
-    
-    touchEnd: function () {
-      this.toggleProperty('isShowingBody')
-      this.toggleProperty('clicked')
-    }
-  }
-});
