--- app/templates/components/filter-nav.hbs
@@ -1,10 +1,0 @@
-<div class="nav-3">
-  <div class="filter-by" {{action "touchEnd" }}>
-    <span>Filter by</span>
-    <p>All Employees</p>
-    <img src="/assets/images/accordion-open.svg" alt="img" class='open {{if clicked "turn"}}'>
-  </div>
-</div>
-
-
-{{filter-by isShowingBody=isShowingBody usersList=usersList}}
